+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - The `spec::Parameter::schema` field is now of type `ObjectOrReference<Schema>`.
+- Add `Operation::extensions` field.
 - Minimum supported Rust version (MSRV) is now 1.70.
 
 ## 0.8.1